Ensure namespace and SA are created (#168)
The newly created namespace and service account are not available for
login and switching the context immediatly. There was an explicit sleep
introduced in PR #168 as a workaround to wait for resource availability.
This PR introduces an incremental check for resource availability. Once
all the resources are realized, the program exit normally.

* Suggest to re-run the job if there is any infra failure